Modified a number of models by adding more member functions to refine manipulation.  Any information that was being added by overwriting .save()'s has been transfered into .init() methods.  Generally speaking if a model has a .init() method, it will perform similarly to .save(), but also save the event coresponding to creating an instance of that model.
All events being triggered by signals have been scrapped as well, the intuition being that if you're not using the provided method to manipulate a model instance - and instead are manipulating the model directly (e.g. segment.tags.add() vs segment.tag()) - the event associated with that event *shouldn't* fire.  This will also come in usefull later on if we want to associate other functionality within a method.  For example, if we wanted to do something in the background every time a segment was tagged, it be better to add to a method than try and hook into tag event being fired. <- this is a simplification, but you get the idea

Expanded the event system.